model/requirements.txt
streamlit
pillow
pytesseract
pdf2image
pymupdf
requests
